Configurable Environment Variables Added a FromStr implementation to debian::MultiarchName Implment functionality to load the environment variables associate with packages from the project.toml file use PACKAGE_ENV_VARS constand to map default env vars to be defined if the package is in the project.toml file - these can be overwritten by the values in the project.toml file define the package_env_vars before the env vars in the project.toml file update README to reflect the PACKAGE_ENV_VARS
testing the buildpack, added a variable for skipped_packages.  Will incorporate that into install_packages

move the env logic into configure_layer_environment.  Iterate through the skipped_packages to make sure all the env vars are set and the post install command are run

flushed out a test in order to accomindate the new parameters passed to configure_layer_environment

finished the test for skipped packages

update the unit test to use a fixture to create the env correctly

use  a fixture for the test project.toml file; update the README